new organisation of genre  db functions
